Sequencial isotope annotation algorithm implementation.
TODO: split the candidate maker from the isotope annotator and code the adduct annotation and the judge
