+++ NAMESPACE
@@ -1,5 +1,5 @@
 # Generated by roxygen2 but modified to handle correctly rccp, add here required methods
-importFrom(Rcpp,evalCpp) 
+importFrom(Rcpp,evalCpp)  
 useDynLib(rMSIproc) 
 export(Smoothing)
 export(NoiseEstimationFFTCosWin)
@@ -7,6 +7,7 @@ export(NoiseEstimationFFTExpWin)
 export(NoiseEstimationFFTCosWinMat)
 export(NoiseEstimationFFTExpWinMat)
 export(DetectPeaks)
+export(Deisotoper)
 export(printrMSIdataInfo)
 export(loadDataCube)
 export(ProcessImage)
@@ -25,3 +26,5 @@ export(plotPeakImage)
 export(plotClusterImage)
 export(plotValuesImage)
 export(AverageSpectrum)
+export(export_imzMLpeakList)
+export(import_imzMLpeakList)
