+++ R/RcppExports.R
@@ -1,8 +1,8 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-PeakSelectorC <- function(massChannels, numPixels, NumIso, PeakMtx, massVec, ppmT) {
-    .Call('_rMSIproc_PeakSelectorC', PACKAGE = 'rMSIproc', massChannels, numPixels, NumIso, PeakMtx, massVec, ppmT)
+IsotopeAnnotator <- function(massPeaks, massChannels, numPixels, NumIso, PeakMtx, massVec, massChanVec, scansT, accu) {
+    .Call('_rMSIproc_IsotopeAnnotator', PACKAGE = 'rMSIproc', massPeaks, massChannels, numPixels, NumIso, PeakMtx, massVec, massChanVec, scansT, accu)
 }
 
 AlignSpectrumToReference <- function(mass, ref, x, bilinear = FALSE, lagRefLow = 0.1, lagRefMid = 0.5, lagRefHigh = 0.9, iterations = 1L, lagLimitppm = 200, fftOverSampling = 2L) {
@@ -25,12 +25,8 @@ FullImageAlign <- function(fileNames, mass, refSpectrum, numRows, dataType, numO
     .Call('_rMSIproc_FullImageAlign', PACKAGE = 'rMSIproc', fileNames, mass, refSpectrum, numRows, dataType, numOfThreads, AlignmentBilinear, AlignmentIterations, AlignmentMaxShiftPpm, RefLow, RefMid, RefHigh, OverSampling)
 }
 
-FullImagePeakPicking <- function(fileNames, mass, numRows, dataType, numOfThreads, SNR = 5, WinSize = 10L, InterpolationUpSampling = 10L, doBinning = TRUE, binningTolerance = 100, binningFilter = 0.9, binningIn_ppm = TRUE) {
-    .Call('_rMSIproc_FullImagePeakPicking', PACKAGE = 'rMSIproc', fileNames, mass, numRows, dataType, numOfThreads, SNR, WinSize, InterpolationUpSampling, doBinning, binningTolerance, binningFilter, binningIn_ppm)
-}
-
-MergePeakMatricesC <- function(PeakMatrices, binningTolerance = 100, binningFilter = 0.01) {
-    .Call('_rMSIproc_MergePeakMatricesC', PACKAGE = 'rMSIproc', PeakMatrices, binningTolerance, binningFilter)
+FullImagePeakPicking <- function(fileNames, mass, numRows, dataType, numOfThreads, SNR = 5, WinSize = 10L, InterpolationUpSampling = 10L, doBinning = TRUE, binningTolerance = 100, binningFilter = 0.9, binningIn_ppm = TRUE, exportPeakList = FALSE) {
+    .Call('_rMSIproc_FullImagePeakPicking', PACKAGE = 'rMSIproc', fileNames, mass, numRows, dataType, numOfThreads, SNR, WinSize, InterpolationUpSampling, doBinning, binningTolerance, binningFilter, binningIn_ppm, exportPeakList)
 }
 
 ReplacePeakMatrixZeros <- function(PeakMatrix, fileNames, mass, numRows, dataType, numOfThreads, WinSize = 10L, InterpolationUpSampling = 10L) {
@@ -87,6 +83,10 @@ NoiseEstimationFFTCosWinMat <- function(x, filWinSize = 40L) {
 #' @return a
 NoiseEstimationFFTExpWinMat <- function(x, filWinSize = 40L) {
     .Call('_rMSIproc_NoiseEstimationFFTExpWinMat', PACKAGE = 'rMSIproc', x, filWinSize)
+}
+
+MergePeakMatricesC <- function(PeakMatrices, binningTolerance = 100, binningFilter = 0.01) {
+    .Call('_rMSIproc_MergePeakMatricesC', PACKAGE = 'rMSIproc', PeakMatrices, binningTolerance, binningFilter)
 }
 
 #'LoadPeakMatrix.
