+++ R/mass_calibration.R
@@ -28,7 +28,7 @@
 #'
 CalibrationWindow<-function( mass, intensity, peak_win_size = 20, win_title = "", CalibrationSpan = 0.75, NotPerformCalibration = F)
 {
-  options(guiToolkit="RGtk2") # Força que toolquit sigu GTK pq fas crides directes a events GTK!!!
+  options(guiToolkit="RGtk2") # Fora que toolquit sigu GTK pq fas crides directes a events GTK!!!
   oldWarning<-options()$warn
   options(warn = -1)
