+++ man/ProcessImage.Rd
@@ -6,14 +6,27 @@
 \usage{
 ProcessImage(img, EnableSmoothing = T, SmoothingKernelSize = 5,
   EnableAlignment = T, AlignmentIterations = 3,
+<<<<<<< HEAD
+  AlignmentMaxShiftppm = 200, AlignmentBilinear = F,
+  AlignmentRefLow = 0, AlignmentRefMid = 0.5, AlignmentRefHigh = 1,
+  AlignmentOversampling = 2, EnableCalibration = T,
+  CalibrationPeakWin = 20, EnablePeakPicking = T, SNR = 5,
+  peakWindow = 10, peakUpSampling = 10, UseBinning = T,
+  BinTolerance = 100, BinFilter = 0.05, BinToleranceUsingPPM = T,
+  EnableSpectraNormalization = T, EnableTICNorm = T,
+  EnableRMSNorm = T, EnableMAXNorm = T, EnableTICAcqNorm = T,
+  NumOfThreads = parallel::detectCores(), CalSpan = 0.75)
+=======
   AlignmentMaxShiftppm = 200, AlignmentBilinear = F, AlignmentRefLow = 0,
   AlignmentRefMid = 0.5, AlignmentRefHigh = 1, AlignmentOversampling = 2,
   EnableCalibration = T, CalibrationPeakWin = 20, EnablePeakPicking = T,
   SNR = 5, peakWindow = 10, peakUpSampling = 10, UseBinning = T,
-  BinTolerance = 100, BinFilter = 0.05, BinToleranceUsingPPM = T,
+  BinTolerance = 5, BinFilter = 0.05, BinToleranceUsingPPM = F,
   EnableSpectraNormalization = T, EnableTICNorm = T, EnableRMSNorm = T,
   EnableMAXNorm = T, EnableTICAcqNorm = T,
-  NumOfThreads = parallel::detectCores(), CalSpan = 0.75)
+  NumOfThreads = parallel::detectCores(), CalSpan = 0.75,
+  ExportPeakList = F)
+>>>>>>> 3bff3ad101554c83c23b21e026537a21a5bc9ecc
 }
 \arguments{
 \item{img}{an rMSI data object to process or a list of rMSI objects if various datasets must merged for processing.}
@@ -69,6 +82,8 @@ ProcessImage(img, EnableSmoothing = T, SmoothingKernelSize = 5,
 \item{NumOfThreads}{the number number of threads used to process the data.}
 
 \item{CalSpan}{the used span for the loess fittin used in mass calibration.}
+
+\item{ExportPeakList}{a boolean detailing wheter the un-binned peak list must be exported or not.}
 }
 \value{
 a named list containing:
