+++ man/plotValuesImage.Rd
@@ -1,19 +1,19 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/plotPeakMatrix.R
-\name{plotClusterImage}
-\alias{plotClusterImage}
-\title{plotClusterImage.}
+\name{plotValuesImage}
+\alias{plotValuesImage}
+\title{plotValuesImage}
 \usage{
-plotClusterImage(peakMatrix, clusters, nrow = 2, ncol = 2, byrow = T,
+plotValuesImage(peakMatrix, values, nrow = 2, ncol = 2, byrow = T,
   margin = 20, img_names = peakMatrix$names, labels = img_names,
   rotations = rep(0, length(img_names)), mirror_x = rep(F,
   length(img_names)), mirror_y = rep(F, length(img_names)),
-  pixel_size_um = 100)
+  pixel_size_um = 100, scale_title = "", light = 8)
 }
 \arguments{
 \item{peakMatrix}{the peak matrix in an rMSIproc object.}
 
-\item{clusters}{a vector with integer number according the cluster of each pixel.}
+\item{values}{a vector with the values to be plotted in each pixel.}
 
 \item{nrow}{number of rows of the plotted matrix layout (only used if byrow == F)}
 
@@ -34,10 +34,11 @@ plotClusterImage(peakMatrix, clusters, nrow = 2, ncol = 2, byrow = T,
 \item{mirror_y}{a bool vector specifing if each image must be flipped or not in Y direction prior to rotation.}
 
 \item{pixel_size_um}{the pixel resolution in um.}
-}
-\value{
-a vector with the used color for each cluster sorted according clustering numering in assending order.
+
+\item{scale_title}{the label for the color scale.}
+
+\item{light}{the lighting of the plotted image.}
 }
 \description{
-Plot a segmentation image with the user-given clusters.
+Plot arbitrary values in a MS images raster positions.
 }
