+++ src/annotationcandidates.h
@@ -1,0 +1,42 @@
+/*************************************************************************
+ * 
+ *     Copyright (C) 2018 Lluc Sementé Fernández
+ * 
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ * 
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ * 
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ **************************************************************************/
+
+// #ifndef DEISOTOPER_H
+// #define DEISOTOPER_H
+// 
+// #include <Rcpp.h>
+// #include <stdlib.h>
+// #include <cmath>
+// 
+// using namespace Rcpp;
+// 
+// class AnnotationCandidates
+// {
+// public:
+//   
+//   typedef struct 
+//   {
+//     int 
+//     
+//   }AnnDef;
+//   
+// private:
+//   double *CandidatesDelay; //Number of candidates for each index (length = mass peaks)
+//   double *CandidatesVector; //Candidates for each peak  
+// 
+// };
