+++ src/mtpeakpicking.cpp
@@ -21,18 +21,8 @@
 #include <vector>
 #include <limits>
 #include "mtpeakpicking.h"
+#include "peakbinning.h"
 using namespace Rcpp;
-
-MTPeakPicking::MTPeakPicking():
-  ThreadingMsiProc()
-{
-  //Just init dummy objects to allow the desctructor to work properly
-  numOfThreadsDouble = 0;
-  numOfPixels = 0;
-  binSizeInppm = true;
-  peakObj = new PeakPicking*[1];
-  mPeaks =  new PeakPicking::Peaks*[1];
-}
 
 MTPeakPicking::MTPeakPicking(ImgProcDef imgRunInfo) : 
   ThreadingMsiProc( imgRunInfo.numOfThreads, false, imgRunInfo.fileNames, imgRunInfo.massChannels, imgRunInfo.numRows, imgRunInfo.dataType )
@@ -44,16 +34,12 @@ MTPeakPicking::MTPeakPicking(ImgProcDef imgRunInfo) :
   }
   
   minSNR = imgRunInfo.SNR;
-  binSize = imgRunInfo.tolerance;
-  binSizeInppm = imgRunInfo.tolerance_in_ppm;
-  binFilter = imgRunInfo.filter;
   numOfPixels = 0;
   for( int i = 0; i < imgRunInfo.fileNames.length(); i++)
   {
     numOfPixels +=  imgRunInfo.numRows[i];
   }
   mPeaks =  new PeakPicking::Peaks*[numOfPixels];
-  bDoBinning = imgRunInfo.performBinning;
 }
 
 MTPeakPicking::~MTPeakPicking()
@@ -71,290 +57,32 @@ MTPeakPicking::~MTPeakPicking()
   delete[] mPeaks;
 }
 
-List MTPeakPicking::Run()
+void MTPeakPicking::Run()
 {
   //Run peak-picking in mutli-threading
   runMSIProcessingCpp();
-
-  //Binning
-  if(bDoBinning)
-  {
-    return BinPeaks();
-  }
-  else
-  {
-    List pkLst(numOfPixels);
-    NumericVector pkMass, pkIntensity, pkSNR, pkArea, pkBinSize;
-    for(int i = 0; i < numOfPixels; i++)
-    {
-      Rcout<<"Copying peaks "<< (i+1) <<" of "<< numOfPixels <<"\n";
-      pkMass = NumericVector(mPeaks[i]->mass.size());
-      pkIntensity = NumericVector(mPeaks[i]->intensity.size());
-      pkSNR = NumericVector(mPeaks[i]->SNR.size());
-      pkArea = NumericVector(mPeaks[i]->area.size());
-      pkBinSize = NumericVector(mPeaks[i]->binSize.size());
-      memcpy(pkMass.begin(), mPeaks[i]->mass.data(), sizeof(double)*mPeaks[i]->mass.size());
-      memcpy(pkIntensity.begin(), mPeaks[i]->intensity.data(), sizeof(double)*mPeaks[i]->intensity.size());
-      memcpy(pkSNR.begin(), mPeaks[i]->SNR.data(), sizeof(double)*mPeaks[i]->SNR.size());
-      memcpy(pkArea.begin(), mPeaks[i]->area.data(), sizeof(double)*mPeaks[i]->area.size());
-      memcpy(pkBinSize.begin(), mPeaks[i]->binSize.data(), sizeof(double)*mPeaks[i]->binSize.size());
-      pkLst[i] = List::create( Named("mass") = pkMass, Named("intensity") = pkIntensity, Named("SNR") = pkSNR, Named("area") = pkArea, Named("binSize") = pkBinSize);
-    }
-    return pkLst;
-  }
 }
 
-void MTPeakPicking::AppendPeaksAsMatrix(List peaksLst)
+Rcpp::List MTPeakPicking::ExportPeakList()
 {
-  //Create a local Peaks to handle all peaks
-  int localNumPixels = numOfPixels + (as<NumericMatrix>(peaksLst["intensity"])).rows();
-  PeakPicking::Peaks **localPeaks =  new PeakPicking::Peaks*[localNumPixels];
-  
-  //Copy the current Peaks to the local Peaks object 
-  for( int i = 0; i < numOfPixels; i++)
-  {
-    localPeaks[i] = mPeaks[i];
-  }
-  delete[] mPeaks; //Delete the old references array, but keeps data pointers in localPeaks
-  mPeaks = localPeaks; //Move to localPeaks pointer and set it as current peaks pointer
-  
-  //Append new peaks 
-  for( int i = 0; i < (localNumPixels - numOfPixels); i++)
-  {
-    mPeaks[ i + numOfPixels] = new PeakPicking::Peaks();
-    
-    for(int j = 0; j < (as<NumericVector>(peaksLst["mass"])).length(); j++)
-    {
-      if( (as<NumericMatrix>(peaksLst["intensity"]))(i, j) > 0  )
-      {
-        mPeaks[ i + numOfPixels]->mass.push_back( (as<NumericVector>(peaksLst["mass"]))(j) ); 
-        mPeaks[ i + numOfPixels]->intensity.push_back( (as<NumericMatrix>(peaksLst["intensity"]))(i, j) );
-        mPeaks[ i + numOfPixels]->SNR.push_back( (as<NumericMatrix>(peaksLst["SNR"]))(i, j)  );   
-        mPeaks[ i + numOfPixels]->area.push_back( (as<NumericMatrix>(peaksLst["area"]))(i, j)  );
-      }
-    }
-  }
-  numOfPixels = localNumPixels;
-}
-  
-List MTPeakPicking::BinPeaks()
-{
-  //Compute each peak obj TIC, I'll start binning from the highest spectrum which is assumed to have a lower mass error (at least for higher peaks)
-  double Tic[numOfPixels];
+  List pkLst(numOfPixels);
+  NumericVector pkMass, pkIntensity, pkSNR, pkArea, pkBinSize;
   for(int i = 0; i < numOfPixels; i++)
   {
-    Tic[i] = 0.0;
-    for( unsigned int j = 0; j < mPeaks[i]->intensity.size(); j++)
-    {
-      Tic[i] += mPeaks[i]->intensity[j];
-    }
+    Rcout<<"Copying peaks "<< (i+1) <<" of "<< numOfPixels <<"\n";
+    pkMass = NumericVector(mPeaks[i]->mass.size());
+    pkIntensity = NumericVector(mPeaks[i]->intensity.size());
+    pkSNR = NumericVector(mPeaks[i]->SNR.size());
+    pkArea = NumericVector(mPeaks[i]->area.size());
+    pkBinSize = NumericVector(mPeaks[i]->binSize.size());
+    memcpy(pkMass.begin(), mPeaks[i]->mass.data(), sizeof(double)*mPeaks[i]->mass.size());
+    memcpy(pkIntensity.begin(), mPeaks[i]->intensity.data(), sizeof(double)*mPeaks[i]->intensity.size());
+    memcpy(pkSNR.begin(), mPeaks[i]->SNR.data(), sizeof(double)*mPeaks[i]->SNR.size());
+    memcpy(pkArea.begin(), mPeaks[i]->area.data(), sizeof(double)*mPeaks[i]->area.size());
+    memcpy(pkBinSize.begin(), mPeaks[i]->binSize.data(), sizeof(double)*mPeaks[i]->binSize.size());
+    pkLst[i] = List::create( Named("mass") = pkMass, Named("intensity") = pkIntensity, Named("SNR") = pkSNR, Named("area") = pkArea, Named("binSize") = pkBinSize);
   }
-  
-  //Apply binning starting with most intens spectra
-  double dMax;
-  int iMax;
-  std::vector<std::vector<TBin> > binMat; //A matrix containing binning values
-  NumericVector binMass; //The mass name for each matrix column
-  NumericVector binSizeVec; //The raw spectra bin size for each matrix column
-  NumericVector binAvgInt; //The average spectra intensity of each bin
-  Rcout<<"Binning...\n";
-  int iCount = 0;
-
-  double alpha = 0; //Use to calculate bin mass related to peak intensity
-  while(true)
-  {
-    Rcout<<iCount<<"/"<<numOfPixels<<"\n";
-    iCount++;
-    //Locate current most intens spectrum
-    iMax = -1;
-    dMax = -1.0;
-    for(int i = 0; i < numOfPixels; i++)
-    {
-      if(Tic[i] > dMax && Tic[i] >= 0)
-      {
-        dMax = Tic[i];
-        iMax = i;
-      }
-    }
-    
-    //End condition
-    if(iMax == -1)
-    {
-      break;
-    }
-    
-    Tic[iMax] = -1.0; //Mark current spectrum as processed by deleting its TIC
-
-    while( mPeaks[iMax]->mass.size() > 0 )
-    {
-      binMass.push_back(mPeaks[iMax]->mass[0]); //Append element to the mass vector (names of bin Matrix)
-      binSizeVec.push_back(mPeaks[iMax]->binSize[0]); //Append element to the binSize vector (names of bin Matrix)
-      binAvgInt.push_back(mPeaks[iMax]->intensity[0]); //Append element to intensity bin
-      binMat.resize(binMat.size() + 1); //Append new column
-      binMat[binMat.size() - 1].resize(numOfPixels); //Extenend all new column elements
-      binMat[binMat.size() - 1][iMax].intensity = mPeaks[iMax]->intensity[0]; 
-      binMat[binMat.size() - 1][iMax].SNR = mPeaks[iMax]->SNR[0];
-      binMat[binMat.size() - 1][iMax].area = mPeaks[iMax]->area[0];
-      
-      //Delete current peak
-      mPeaks[iMax]->mass.erase(mPeaks[iMax]->mass.begin());
-      mPeaks[iMax]->intensity.erase(mPeaks[iMax]->intensity.begin());
-      mPeaks[iMax]->SNR.erase(mPeaks[iMax]->SNR.begin());
-      mPeaks[iMax]->area.erase(mPeaks[iMax]->area.begin());
-      mPeaks[iMax]->binSize.erase(mPeaks[iMax]->binSize.begin());
-      
-      double numMasses = 1.0; //number of peak masses used to compute each mass centroid
-      int countPeaks = 1; //Number of peaks in current column
-      
-      for( int j = 0; j < numOfPixels; j++)
-      {
-        if(j != iMax) //Do not process current Peaks obj
-        {
-          //Find the nearest peak, the nearest peak postion is reatined in iPos var
-          double dist = 0.0;
-          double dist_ppm;
-          double dist_ant;
-          double dist_ant_ppm = 1e50;
-          int iPos = 0;
-          for( unsigned int imass = 0; imass <  mPeaks[j]->mass.size(); imass++)
-          {
-            iPos = imass;
-            dist = binMass[binMass.size() - 1] - mPeaks[j]->mass[imass];
-            dist_ppm = 1e6*dist/binMass[binMass.size() - 1]; //Translation to ppm
-            if(dist_ppm <= 0)
-            {
-              if(std::abs(dist_ant_ppm) < std::abs(dist_ppm))
-              {
-                iPos = imass - 1;
-                dist_ppm = dist_ant_ppm;
-                dist = dist_ant;
-              }
-              break; //avoid all the loop
-            }
-            dist_ant_ppm = dist_ppm;
-            dist_ant = dist;
-          }
-          
-          //Select the kind of binning tolerance
-          double compTolerance;
-          if(binSizeInppm)
-          {
-            dist = dist_ppm;
-            compTolerance = binSize;
-          }
-          else
-          {
-            compTolerance = binSize * binSizeVec[binSizeVec.size() - 1];
-          } 
-          
-          //Check if is in the same mass bin and fill the matrix value accordingly
-          if(std::abs(dist) <= compTolerance && mPeaks[j]->mass.size() > 0)
-          {
-            binMat[binMat.size() - 1][j].intensity = mPeaks[j]->intensity[iPos];
-            binMat[binMat.size() - 1][j].SNR = mPeaks[j]->SNR[iPos];
-            binMat[binMat.size() - 1][j].area = mPeaks[j]->area[iPos];
-            if(mPeaks[j]->SNR[iPos] >= minSNR)
-            {
-              //Count peak only if its SNR fits the min SNR value
-              countPeaks++;
-            }
-            
-            //Recompute mass centroid using a continuous average
-            alpha =  binAvgInt[binAvgInt.size() - 1] / ( binAvgInt[binAvgInt.size() - 1] + mPeaks[j]->intensity[iPos]);
-            binMass[binMass.size() - 1] = alpha * binMass[binMass.size() - 1] +  (1.0 - alpha) * mPeaks[j]->mass[iPos];
-            binSizeVec[binSizeVec.size() - 1] *= numMasses;
-            binAvgInt[binAvgInt.size() - 1] *= numMasses;
-            binSizeVec[binSizeVec.size() - 1] +=  mPeaks[j]->binSize[iPos];
-            binAvgInt[binAvgInt.size() - 1] += mPeaks[j]->intensity[iPos];
-            numMasses++;
-            binSizeVec[binSizeVec.size() - 1] /= numMasses;
-            binAvgInt[binAvgInt.size() - 1] /= numMasses;
-            
-            //Delete datapoint from current peaks
-            mPeaks[j]->mass.erase(mPeaks[j]->mass.begin() + iPos);
-            mPeaks[j]->intensity.erase(mPeaks[j]->intensity.begin() + iPos);
-            mPeaks[j]->SNR.erase(mPeaks[j]->SNR.begin() + iPos);
-            mPeaks[j]->area.erase(mPeaks[j]->area.begin() + iPos);
-            mPeaks[j]->binSize.erase(mPeaks[j]->binSize.begin() + iPos);
-          }
-          else
-          {
-            binMat[binMat.size() - 1][j].intensity = 0.0;
-            binMat[binMat.size() - 1][j].SNR = 0.0;
-            binMat[binMat.size() - 1][j].area = 0.0;
-          }
-        }
-      }
-      
-      //Delete last column if it does not fit minimum filter criterion
-      if(countPeaks < (int)(numOfPixels * binFilter))
-      {
-        binMass.erase(binMass.begin() + binMass.size() - 1);
-        binSizeVec.erase(binSizeVec.begin() + binSizeVec.size() - 1);
-        binAvgInt.erase(binAvgInt.begin() + binAvgInt.size() - 1);
-        binMat.erase(binMat.begin() + binMat.size() - 1);
-      }
-      
-    }
-  }
-  Rcout<<"Bining complete with a total number of "<<binMass.size()<<" bins\n";
-  
-  //Copy data to R matrices
-  Rcout<<"Coping data to R object...\n";
-  NumericMatrix binMatIntensity(numOfPixels, binMass.size());
-  NumericMatrix binMatSNR(numOfPixels, binMass.size());
-  NumericMatrix binMatArea(numOfPixels, binMass.size());
-  
-  //Sort columns by mass
-  Rcout<<"Sorting columns by mass...\n";
-  NumericVector massSorting(binMass.size());
-  NumericVector binsizeSorting(binSizeVec.size());
-  memcpy(massSorting.begin(), binMass.begin(), sizeof(double)*binMass.size());
-  int sortedInds[binMass.size()];
-  double minVal;
-  for(int i = 0; i < binMass.size(); i++)
-  {
-    minVal = std::numeric_limits<double>::max();
-    for( int j = 0; j < binMass.size(); j++ )
-    {
-      if( massSorting[j] < minVal && massSorting[j] > 0 )
-      {
-        minVal = massSorting[j];
-        sortedInds[i] = j;
-      }  
-    }
-    massSorting[ sortedInds[i]  ] = -1; //Mark as sorted
-  }
-  
-  //Copy the mass axis sorting it
-  for(int i = 0; i < binMass.size(); i++)
-  {
-    massSorting[i] = binMass[sortedInds[i]];
-    binsizeSorting[i] = binSizeVec[sortedInds[i]];
-  }
-  
-  //Copy the matrix sorting it
-  for( int ir = 0;  ir < numOfPixels; ir++)
-  {
-    for(int ic = 0; ic < binMass.size(); ic++)
-    {
-      binMatIntensity(ir, ic ) = binMat[sortedInds[ic]][ir].intensity;
-      binMatSNR      (ir, ic ) = binMat[sortedInds[ic]][ir].SNR;
-      binMatArea     (ir, ic ) = binMat[sortedInds[ic]][ir].area;
-    }
-  }
-  
-  return List::create( Named("mass") = massSorting, Named("intensity") = binMatIntensity, Named("SNR") = binMatSNR, Named("area") = binMatArea, Named("binsize") = binsizeSorting );
-}
-
-void MTPeakPicking::SetBinSize(double value)
-{
-  binSize = value;
-}
-
-void MTPeakPicking::SetBinFilter(double value)
-{
-  binFilter = value;
+  return pkLst;
 }
 
 void MTPeakPicking::ProcessingFunction(int threadSlot)
@@ -369,12 +97,18 @@ void MTPeakPicking::ProcessingFunction(int threadSlot)
   }
 }
 
+PeakPicking::Peaks **MTPeakPicking::getPeakList()
+{
+  return mPeaks;
+}
+
 // [[Rcpp::export]]
 List FullImagePeakPicking ( StringVector fileNames, 
                                 NumericVector mass, IntegerVector numRows,
                                 String dataType, int numOfThreads, 
                                 double SNR = 5, int WinSize = 10, int InterpolationUpSampling = 10, 
-                                bool doBinning = true, double binningTolerance = 100, double binningFilter = 0.9, bool binningIn_ppm = true)
+                                bool doBinning = true, double binningTolerance = 100, double binningFilter = 0.9, bool binningIn_ppm = true,
+                                bool exportPeakList = false)
 {
   //Copy R data to C arrays
   double *massC = new double[mass.length()];
@@ -392,27 +126,33 @@ List FullImagePeakPicking ( StringVector fileNames,
   myProcParams.peakInterpolationUpSampling  = InterpolationUpSampling;
   myProcParams.peakWinSize = WinSize;
   myProcParams.SNR = SNR;
-  myProcParams.tolerance = binningTolerance;
-  myProcParams.filter = binningFilter;
-  myProcParams.performBinning = doBinning;
-  myProcParams.tolerance_in_ppm = binningIn_ppm;
   
   MTPeakPicking myPeakPicking(myProcParams);
   delete[] massC;
   
-  return myPeakPicking.Run();
-}
-
-// [[Rcpp::export]]
-List MergePeakMatricesC( List PeakMatrices, double binningTolerance = 100, double binningFilter = 0.01 )
-{
-  MTPeakPicking myPeakPicking;
-  for( int i = 0; i < PeakMatrices.length(); i++)
+  myPeakPicking.Run();
+  PeakPicking::Peaks **CpeakList_ptr = myPeakPicking.getPeakList();
+  
+  //Export the peak list without binning for imzML centroid convertion, this must be called before the binning since binning will destroy the peak list
+  List peakList;
+  if( exportPeakList || !doBinning )
   {
-    Rcout<<"Merging peak matrix "<<(i+1)<<" of "<<PeakMatrices.length()<<"\n";
-    myPeakPicking.AppendPeaksAsMatrix(PeakMatrices[i]);  
+    peakList = myPeakPicking.ExportPeakList();
+  } 
+    
+  //Call binning if it is enabled
+  List peakMat;
+  if(doBinning)
+  {
+    int pixelCount = 0;
+    for( int i = 0; i < fileNames.length(); i++)
+    {
+      pixelCount +=  numRowsC[i];
+    }
+    PeakBinning myPeakBinning(CpeakList_ptr, pixelCount, binningTolerance, binningIn_ppm, binningFilter);
+    peakMat = myPeakBinning.BinPeaks(); 
   }
-  myPeakPicking.SetBinFilter(binningFilter);
-  myPeakPicking.SetBinSize(binningTolerance);
-  return myPeakPicking.BinPeaks();
+  
+  //The peak list object is automatically destroyed by the MTPeakPicking destructor
+  return List::create( Named("PeakMatrix") = peakMat, Named("PeakList") = peakList);
 }
